This is a holy grail of a find.
This will pull the currently bound key as a string from a command.
Use case - allow users to rebind keys for a command, this can then pull what they have bound as a string. Very useful for ui/helptext/notifications to display the correct key in use.

The codes messey but its because the special chars and is needed. We need this though sooooo yea.