--Upgrade skills
Added an in house skill system, nests skills tables inside player metadata. This is convenient for us as data tied to player - no save logic/load logic required on our end. Tested in qbx/esx no issues. Given slapping *large* tables in this would be bad but this is tiny and useful.