experimental but good
Early on I gave the name management to banking, I know I know it was dumb. BUT we can either deprecate them and move this to it or keep boss menu as its own module and deprecate the banking in time to a name that makes sense. Open to suggestions here